move types to custom types file
