+++ custom_types.py
@@ -1,8 +1,9 @@
 from dataclasses import dataclass
 from enum import Enum
-from typing import Optional
+from typing import Literal, Optional
+from jaxtyping import Int
+from torch import Tensor
 
-from buffer.base_buff import IOPair
 
 """
 Task type to list of tuples where tuple[0] is the model answer and tuple[1] is the ground truth answer
@@ -31,3 +32,20 @@ class TaskType(Enum):
 class Role(Enum):
     SOLVER = "SOLVER"
     PROPOSER = "PROPOSER"
+
+@dataclass
+class IOPair:
+    input_str: str
+    output_str: str
+
+
+@dataclass
+class Sample:
+    snippet: str
+    function_io: list[IOPair]
+    input_types: Literal["str", "int", "list", "tuple"]
+    output_types: Literal["str", "int", "list", "tuple"]
+    message: str
+    imports: str  # executable string
+    prompt_tokens: list[str]
+    sample_ids: Int[Tensor, "seq_len"]
